created a db object for handling the database, and destroys the connection when necessary.
